Huge update.
Occurrences of NOP vs nop:

Even if lower case nop proceeds through the decode/execute stage, nothing can happen because the instruction has no values for any of its fields. It cannot read register values in the decode stage and it cannot perform any tasks in the execute stage.
• If the first instruction in the list is NOP for the decode stage the stage is exited immediately.
• If the first instruction in the list is NOP for the execute stage the stage is exited immediately.

This may be why NOP is checked as the second instruction in the WB stage but nop is not checked
• In the instruction fetch stage, NOP is fed into the pipeline as the second instruction if a pair cannot be found for entering the pipeline

• NOP is inserted for four instructions of two stages in the case of a branch misprediction in simulator.cpp

It does not make sense to increment the instruction count by 2 if one of the instructions is a nop (even if it is lower case)
• NOP determines if the increment is 1 in the write back stage

The first instruction of instruction list of the pipeline stages was before the second instruction of the instruction list, even if they proceed through the pipeline in the same cycle.

Is it possible for currentInstructionList[0] to not be a nop, but currentInstructionList[1] to be a nop? For capital NOP, this is possible due to the fetch stage setting NOP to be the second instruction while the first instruction proceeds through the pipeline.

end, nop, nop, nop is inserted at the end of the instruction list